feat: viewModel으로 sharedPreference 이동 to-do: viewModel에서 sharedPreference만으로 로그인 로그아웃 되도록 구현 후, local 더미데이터를 firebase로 교체
