+++ app/src/main/java/com/android/hanple/utils/ConvertUtils.kt
@@ -15,4 +15,6 @@ object ConvertUtils {
             .atZone(ZoneId.systemDefault()) //안드로이드 OS가 확인한 현재 시간대.
             .toLocalDateTime()
 
+
+    val phoneNumberRegex = "^\\+?\\d{1,4}[- ]?\\d{4,}(?:[- ]?\\d{4,})?\$".toRegex() //SignUpViewModel에 있는 값을 옮김.
 }
