+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.LogInActivity">
+    tools:context=".ui.onboarding.LogInActivity">
 
     <ImageView
         android:id="@+id/img_profile"
