+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:padding="32dp"
     android:background="@color/white"
-    tools:context=".ui.SignUpActivity">
+    tools:context=".ui.onboarding.SignUpActivity">
 
 
     <TextView
